+++ app/src/main/java/com/skeler/scanely/core/actions/ScanActionDetector.kt
@@ -195,6 +195,7 @@ object ScanActionDetector {
             is ScanAction.SendSms -> action.number
             is ScanAction.AddContact -> "${action.name}${action.phone}${action.email}"
             is ScanAction.ShowRaw -> action.text.take(50)
+            is ScanAction.LookupProduct -> action.barcode
         }
     }